The refactoring and test coverage Change list: - the command forming was moved from the AppiumDriver to MobileCommand - MobileCommand is a class now (was an interface). - Improved test coverage. Now there are tests for both Android and iOS (some test were for iOS only)
