+++ src/main/java/io/appium/java_client/NoSuchContextException.java
@@ -18,10 +18,6 @@ package io.appium.java_client;
 
 import org.openqa.selenium.NotFoundException;
 
-/**
- * Thrown by {@link org.openqa.selenium.WebDriver.TargetLocator#context(String) WebDriver.switchTo().context(String
- * name)}.
- */
 @SuppressWarnings("serial")
 public class NoSuchContextException extends NotFoundException {
