+++ src/main/java/io/appium/java_client/pagefactory/OverrideWidget.java
@@ -39,6 +39,8 @@ public @interface OverrideWidget {
      *
      * A declared class should not be abstract. Declared class also should be a subclass
      * of an annotated class/class which is declared by an annotated field.
+     *
+     * @return a class which extends {@link io.appium.java_client.pagefactory.Widget}
      */
     Class<? extends Widget> html() default Widget.class;
     /**
@@ -46,6 +48,8 @@ public @interface OverrideWidget {
      *
      * A declared class should not be abstract. Declared class also should be a subclass
      * of an annotated class/class which is declared by an annotated field.
+     *
+     * @return a class which extends {@link io.appium.java_client.pagefactory.Widget}
      */
     Class<? extends Widget> androidUIAutomator() default Widget.class;
     /**
@@ -53,6 +57,8 @@ public @interface OverrideWidget {
      *
      * A declared class should not be abstract. Declared class also should be a subclass
      * of an annotated class/class which is declared by an annotated field.
+     *
+     * @return a class which extends {@link io.appium.java_client.pagefactory.Widget}
      */
     Class<? extends Widget> iOSUIAutomation() default Widget.class;
     /**
@@ -60,6 +66,8 @@ public @interface OverrideWidget {
      *
      * A declared class should not be abstract. Declared class also should be a subclass
      * of an annotated class/class which is declared by an annotated field.
+     *
+     * @return a class which extends {@link io.appium.java_client.pagefactory.Widget}
      */
     Class<? extends Widget> selendroid() default Widget.class;
 }
